refactor: Notification PendingIntent 생성을 영화 예매 결과 화면이 가지도록 변경
