+++ app/src/main/java/woowacourse/movie/presentation/home/reservation/result/ReservationResultActivity.kt
@@ -1,11 +1,13 @@
 package woowacourse.movie.presentation.home.reservation.result
 
+import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import woowacourse.movie.R
 import woowacourse.movie.databinding.FragmentReservationResultBinding
 import woowacourse.movie.presentation.MovieTheaterActivity
+import woowacourse.movie.presentation.alarm.AlarmHelper
 import woowacourse.movie.presentation.common.base.BaseActivity
 import woowacourse.movie.presentation.common.extension.getParcelableCompat
 import woowacourse.movie.presentation.common.model.TicketUiModel
@@ -55,5 +57,21 @@ class ReservationResultActivity :
                 BUNDLE_KEY_TICKET,
                 ticket,
             )
+
+        fun newPendingIntent(
+            context: Context,
+            ticket: TicketUiModel,
+        ): PendingIntent {
+            val activityIntent = Intent(context, ReservationResultActivity::class.java)
+            activityIntent.flags = Intent.FLAG_ACTIVITY_CLEAR_TOP
+            activityIntent.putExtra(AlarmHelper.KEY_TICKET, ticket)
+
+            return PendingIntent.getActivity(
+                context,
+                ticket.hashCode(),
+                activityIntent,
+                PendingIntent.FLAG_IMMUTABLE or PendingIntent.FLAG_UPDATE_CURRENT,
+            )
+        }
     }
 }
